+++ examples/project/dialog_box/test_2_input_field.py
@@ -1,12 +1,12 @@
 import pytest
 
-from hardpy import run_dialog_box, set_message
-from hardpy.pytest_hardpy.utils.dialog_box import (
-    DialogBoxWidget,
-    DialogBoxWidgetType,
+from hardpy import (
+    run_dialog_box,
+    set_message,
+    TextInputWidget,
+    NumericInputWidget,
     DialogBox,
 )
-
 
 pytestmark = pytest.mark.module_name("Input field dialog boxes")
 
@@ -16,7 +16,7 @@ def test_text_input():
     dbx = DialogBox(
         dialog_text="Type 'ok' and press the Confirm button",
         title_bar="Example of text input",
-        widget=DialogBoxWidget(DialogBoxWidgetType.TEXT_INPUT),
+        widget=TextInputWidget(),
     )
     response = run_dialog_box(dbx)
     set_message(f"Entered text {response}")
@@ -25,12 +25,12 @@ def test_text_input():
 
 @pytest.mark.case_name("Numeric input")
 def test_num_input():
-    TEST_NUM = 123
+    test_num = 123
     dbx = DialogBox(
-        dialog_text=f"Enter the number {TEST_NUM} and press the Confirm button",
+        dialog_text=f"Enter the number {test_num} and press the Confirm button",
         title_bar="Example of entering a number",
-        widget=DialogBoxWidget(DialogBoxWidgetType.NUMERIC_INPUT),
+        widget=NumericInputWidget(),
     )
     response = int(run_dialog_box(dbx))
     set_message(f"Entered number {response}")
-    assert response == TEST_NUM, "The entered number is not correct"
+    assert response == test_num, "The entered number is not correct"